Correct error handler. Did not work in Chrome
In Chrome, the ‘xhr’ variable had the following value when I hit line
136: "Cannot GET /locales/en-US/translation.json?_=1396024455309”. The
object type of this variable was ‘string’ and it did not have a
‘status’ attribute defined on it.

Since you recently changed this method, changing from ‘error.status’ to
‘xhr.status’, I assume that in some circumstances, the xhr object has a
status attribute. In my case it did not. However, in my case, the
‘status’ parameter DID have a value which I believe you expected the
xhr object to carry in its status attribute. The proposed change allows
for both situations, deferring to the value of the status parameter in
case both are defined.