Default namespace options to expected structure
There are places in code where the namespaces are referenced by `o.ns.namespaces` which throws an error if init has not been run - in particular on `init.addResourceBundle` throws "Cannot call method 'indexOf' of undefined", which is difficult to debug. Setting the default values as the expected structure prevents this difficult-to-debug error.