types: Rework TFunction and translation key suggestions
Allow 'other-ns:some.key' syntax regardless of current namespace(s).
However, we no longer provide these keys via TS autosuggestion, to
not flood the suggestion box with these.
This uses the 'DisableSuggestion' helper type, which unfortunately
allows any capital case prefix to the actual 'other-ns:' prefix.

For getFixedT with multiple namespaces, all the keys in these namespaces
are now suggested without 'ns:' prefix. Using such a prefix is still possible,
though.

Technically, we treat cases with a fixed single namespace and multiple
namespaces the same in `TFuncKey` by flattening an array of namespaces
into a union of namespace strings. Using this we can use distributive
conditional types to handle each namespace individually. See
https://www.typescriptlang.org/docs/handbook/2/conditional-types.html#distributive-conditional-types

For `NormalizeReturn` we use a new `DistributiveKeyOf<T>` helper.
This finds keys that are available in only one/some of the namespaces,
whereas 'keyof T' would only find them if they were in all of the
namespaces.

We simplify and rework the "with options" overload of TFunction.
This no longer allows passing the namespace. That use case is already
covered by other overloads before that. This overload was the reason
for erroneously suggesting unprefixed keys for not currently active
namespaces.

The probably helps with #1887, too.
fixes #1889