Standardize null/undefined checks across codebase
This PR standardizes the pattern of checking for null/undefined values
using `== null` across the codebase. This is a well-known JavaScrip
idion, which was already being used at the codebase on the file
`src/utils.js`.

This simplifies several conditionals while maintaining the same
functionality. It particularly simplifies the check for `defaultNS`, in
which extra care was taken to prevent the case of the variable being
`false`.

Example:
```js
// Before
if (key === undefined || key === null)

// After
if (key == null)
```